+++ src/kernel/arch/i686/ide.h
@@ -2,5 +2,5 @@
 
 #include <stdint.h>
 
-void i686_IDE_Read(uint32_t lba, uint16_t* buffer);
-void i686_IDE_Write(uint32_t lba, const uint16_t* buffer);
+void i686_IDE_Read(uint32_t lba, uint8_t* buffer);
+void i686_IDE_Write(uint32_t lba, const uint8_t* buffer);
